[indexer] Renamings for the FeatureId <-> OsmId mapping.
