+++ generator/cities_ids_builder.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "indexer/fid_bimap.hpp"
+#include "indexer/feature_to_osm.hpp"
 
 #include <string>
 
@@ -9,5 +9,5 @@ namespace generator
 // Takes the FeatureID <-> base::GeoObjectId bidirectional map for features
 // corresponding to cities from |osmToFeaturePath| and serializes it
 // as a section in the mwm file at |dataPath|.
-bool BuildCitiesFidBimap(std::string const & dataPath, std::string const & osmToFeaturePath);
+bool BuildCitiesIds(std::string const & dataPath, std::string const & osmToFeaturePath);
 }  // namespace generator
