+++ indexer/indexer_tests/CMakeLists.txt
@@ -17,9 +17,9 @@ set(
   editable_map_object_test.cpp
   feature_metadata_test.cpp
   feature_names_test.cpp
+  feature_to_osm_tests.cpp
   features_offsets_table_test.cpp
   features_vector_test.cpp
-  fid_bimap_tests.cpp
   index_builder_test.cpp
   interval_index_test.cpp
   locality_index_test.cpp
